--- presentation/src/main/java/com/example/presentation/config/navigation/HomeNavigationHandler.kt
@@ -1,8 +1,0 @@
-package com.example.presentation.config.navigation
-
-interface HomeNavigationHandler {
-    fun navigateToHome()
-    fun navigateToMap()
-    fun navigateToTips()
-    fun navigateToMypage()
-}
