--- presentation/src/main/java/com/example/presentation/config/navigation/TipsNavigationHandler.kt
@@ -1,6 +1,0 @@
-package com.example.presentation.config.navigation
-
-interface TipsNavigationHandler {
-    fun navigateToMagazine()
-    fun navigateToRecipe()
-}
