--- presentation/src/main/java/com/example/presentation/config/navigation/TipsNavigationImpl.kt
@@ -1,14 +1,0 @@
-package com.example.presentation.config.navigation
-
-import androidx.navigation.NavController
-import com.example.presentation.R
-
-class TipsNavigationImpl(private val navController: NavController) : TipsNavigationHandler {
-    override fun navigateToMagazine() {
-        navController.navigate(R.id.homeTipsMagazineFragment)
-    }
-
-    override fun navigateToRecipe() {
-        navController.navigate(R.id.homeTipsRecipeFragment)
-    }
-}
