+++ presentation/src/main/java/com/example/presentation/config/navigation/home/HomeNavigationHandler.kt
@@ -1,0 +1,10 @@
+package com.example.presentation.config.navigation.home
+
+import androidx.navigation.NavController
+
+interface HomeNavigationHandler {
+    fun navigateToHome()
+    fun navigateToMap()
+    fun navigateToTips()
+    fun navigateToMypage()
+}
