+++ presentation/src/main/java/com/example/presentation/config/navigation/main/MainNavigationHandler.kt
@@ -1,0 +1,5 @@
+package com.example.presentation.config.navigation.main
+
+interface MainNavigationHandler {
+    fun navigateToMainHome()
+}
