+++ presentation/src/main/java/com/example/presentation/config/navigation/tips/TipsNavigationHandler.kt
@@ -1,0 +1,8 @@
+package com.example.presentation.config.navigation.tips
+
+import androidx.navigation.NavController
+
+interface TipsNavigationHandler {
+    fun navigateToMagazine()
+    fun navigateToRecipe()
+}
