+++ presentation/src/main/res/layout/activity_main.xml
@@ -18,12 +18,12 @@
             android:layout_height="match_parent">
 
             <androidx.fragment.app.FragmentContainerView
-                android:id="@+id/fcw_home"
-                android:name="androidx.navigation.fragment.NavHostFragment"
+                android:id="@+id/fcw_main_container"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
+                android:name="androidx.navigation.fragment.NavHostFragment"
                 app:defaultNavHost="true"
-                app:navGraph="@navigation/nav_home_graph" />
+                app:navGraph="@navigation/nav_main_graph" />
 
         </androidx.constraintlayout.widget.ConstraintLayout>
