+++ presentation/src/main/res/layout/fragment_main.xml
@@ -13,7 +13,7 @@
         tools:context="view.main.MainFragment">
 
         <androidx.fragment.app.FragmentContainerView
-            android:id="@+id/fcw_main"
+            android:id="@+id/fcw_home"
             android:name="androidx.navigation.fragment.NavHostFragment"
             android:layout_width="match_parent"
             android:layout_height="0dp"
@@ -22,7 +22,7 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
-            app:navGraph="@navigation/nav_main_graph" />
+            app:navGraph="@navigation/nav_home_graph" />
 
         <com.google.android.material.bottomnavigation.BottomNavigationView
             android:id="@+id/bnv_main"
@@ -40,7 +40,7 @@
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/fcw_main"
+            app:layout_constraintTop_toBottomOf="@id/fcw_home"
             app:menu="@menu/menu_bn" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
