+++ presentation/src/main/res/navigation/nav_home_graph.xml
@@ -2,34 +2,34 @@
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/nav_home_graph"
-    app:startDestination="@id/mainFragment">
+    app:startDestination="@id/mainHomeFragment">
 
     <fragment
-        android:id="@+id/mainFragment"
-        android:name="com.example.presentation.view.main.MainFragment"
-        android:label="MainFragment" >
+        android:id="@+id/mainMypageFragment"
+        android:name="com.example.presentation.view.mypage.view.MainMypageFragment"
+        android:label="MainMypageFragment" >
         <action
-            android:id="@+id/action_mainFragment_to_veganTestFragment"
-            app:destination="@id/veganTestFragment" />
+            android:id="@+id/action_mainMypageFragment_to_mainHomeFragment"
+            app:destination="@id/mainHomeFragment" />
     </fragment>
     <fragment
-        android:id="@+id/veganTestFragment"
-        android:name="com.example.presentation.view.home.VeganTestFragment"
-        android:label="VeganTestFragment" >
+        android:id="@+id/veganMapFragment"
+        android:name="com.example.presentation.view.map.view.VeganMapFragment"
+        android:label="VeganMapFragment" >
         <action
-            android:id="@+id/action_veganTestFragment_to_veganTestResultFragment"
-            app:destination="@id/veganTestResultFragment" />
+            android:id="@+id/action_veganMapFragment_to_mainHomeFragment"
+            app:destination="@id/mainHomeFragment" />
     </fragment>
     <fragment
-        android:id="@+id/veganTestResultFragment"
-        android:name="com.example.presentation.view.home.VeganTestResultFragment"
-        android:label="VeganTestResultFragment" >
+        android:id="@+id/tipsFragment"
+        android:name="com.example.presentation.view.tips.view.TipsFragment"
+        android:label="TipsFragment" >
         <action
-            android:id="@+id/action_veganTestResultFragment_to_mainFragment"
-            app:destination="@id/mainFragment" />
+            android:id="@+id/action_tipsFragment_to_mainHomeFragment"
+            app:destination="@id/mainHomeFragment" />
     </fragment>
     <fragment
-        android:id="@+id/mypageEditProfileFragment"
-        android:name="com.example.presentation.view.mypage.view.MypageEditProfileFragment"
-        android:label="MypageEditProfileFragment" />
+        android:id="@+id/mainHomeFragment"
+        android:name="com.example.presentation.view.home.HomeFragment"
+        android:label="MainHomeFragment" />
 </navigation>
